major code refactoring for sadourny! moved all functions into 'sadourny_setup.py' and made 'sadourny_mpi.py', with the serial version to follow soon.
in addition to simply moving several of the functions to the setup file, multiple new wrapper functions were defined:
- creating x points for each process
- creating global objects for process 0
- creating and saving the animation

docstrings for new wrapper functions were written, as well as for set_mpi_bdr, gather_uvh, and wavenum (which could use some more detail).

the indexing of matrices was changed from using 'range(...)' to numpy's notation, which leads to a pretty solid performance boost.