renamed 'flux_ener.f' to be 'flux_ener_components.f', as the Python function that makes use of it calls each subroutine separately. the new fortran file, 'flux_sw_ener.f' calculates all components, flux terms, energy and enstrophy in one go.
created Inds class to be a placeholder for indices.

changed python functions to take params and inds as arguments instead of just params. this is to generalize all functions, as the new fortran function cannot take an array of different types, so one array must hold floats and the other integers (indices).